move `proxy` and `onion` related configurations to their individual modules
